Use the methods with imperial units in `CWeatherDataModel'
because this is the only way QML will be notified properly about the
changes in the corresponding properties. Adding new roles and using the
imperial methods just there didn't work as expected. For some reason the
property values in metric units weren't used in QML at all or were read
at unexpected times and then it wasn't even all properties.