Add positioning support
On app start we now check, if we can get a location from the device's
GPS. If so, we search the nearest available location and retrieve its
weather data from the API. If not, the last position from the settings
is used as fallback. If this value is empty the screen will stay blank
and the busy indicator won't disappear. This will be fixed in the future